Refactor slider and chart logic into reusable components
